코드업 (6039)

+++ python_practice.py
@@ -1,4 +1,5 @@
 a,b = input().split()
 
-c = int(a)**int(b)
+c = float(a)** float(b)
+
 print(c)
